use rust_decimal instead of f32

+++ builder/src/attribute/bonuses/skill.rs
@@ -1,3 +1,5 @@
+use rust_decimal::Decimal;
+
 use crate::{
     attribute::{Attribute, DefaultBonuses, TrackAttribute},
     bonus::{Bonus, BonusType, CloneBonus},
@@ -5,7 +7,7 @@ use crate::{
 };
 
 impl Skill {
-    fn spell_power_bonus(self, sp: SpellPower, value: f32) -> Bonus {
+    fn spell_power_bonus(self, sp: SpellPower, value: Decimal) -> Bonus {
         Bonus::new(
             Attribute::SpellPower(sp),
             BonusType::Stacking,
